+++ cli.go
@@ -3,6 +3,7 @@ package worker
 import (
 	"crypto/rand"
 	"crypto/sha1"
+	"crypto/subtle"
 	"crypto/tls"
 	"crypto/x509"
 	"encoding/hex"
@@ -79,13 +80,6 @@ func NewCLI(c *cli.Context) *CLI {
 // Setup runs one-time preparatory actions and returns a boolean success value
 // that is used to determine if it is safe to invoke the Run func
 func (i *CLI) Setup() (bool, error) {
-	if i.c.String("pprof-port") != "" {
-		// Start net/http/pprof server
-		go func() {
-			http.ListenAndServe(fmt.Sprintf("localhost:%s", i.c.String("pprof-port")), nil)
-		}()
-	}
-
 	if i.c.Bool("debug") {
 		logrus.SetLevel(logrus.DebugLevel)
 	}
@@ -101,6 +95,28 @@ func (i *CLI) Setup() (bool, error) {
 	logrus.SetFormatter(&logrus.TextFormatter{DisableColors: true})
 
 	i.Config = config.FromCLIContext(i.c)
+
+	if i.c.String("pprof-port") != "" && i.c.String("http-api-port") != "" {
+		return false, fmt.Errorf("only one http port is allowed. "+
+			"pprof-port=%v http-api-port=%v",
+			i.c.String("pprof-port"), i.c.String("http-api-port"))
+	}
+	if i.c.String("pprof-port") != "" || i.c.String("http-api-port") != "" {
+		if i.c.String("http-api-auth") != "" {
+			i.setupHTTPAPI()
+		} else {
+			i.logger.Info("skipping HTTP API setup without http-api-auth set")
+		}
+		go func() {
+			httpPort := i.c.String("http-api-port")
+			if httpPort == "" {
+				httpPort = i.c.String("pprof-port")
+			}
+			httpAddr := fmt.Sprintf("localhost:%s", httpPort)
+			i.logger.Info("listening at ", httpAddr)
+			http.ListenAndServe(httpAddr, nil)
+		}()
+	}
 
 	err := i.setupWorkerID()
 	if err != nil {
@@ -385,6 +401,104 @@ func (i *CLI) heartbeatCheck(heartbeatURL, heartbeatAuthToken string) error {
 		i.ProcessorPool.GracefulShutdown(false)
 	}
 	return nil
+}
+
+func (i *CLI) setupHTTPAPI() {
+	i.logger.Info("setting up HTTP API")
+	http.HandleFunc("/worker", i.httpAPI)
+	http.HandleFunc("/worker/", i.httpAPI)
+}
+
+func (i *CLI) httpAPI(w http.ResponseWriter, req *http.Request) {
+	if req.Method == "GET" && req.URL.Path == "/worker" {
+		w.Header().Set("Content-Type", "text/plain;charset=utf-8")
+		w.WriteHeader(http.StatusOK)
+		fmt.Fprintf(w, strings.TrimSpace(`
+Available methods:
+
+- POST /worker/graceful-shutdown
+- POST /worker/graceful-shutdown-pause
+- POST /worker/info
+- POST /worker/pool-decr
+- POST /worker/pool-incr
+- POST /worker/shutdown
+		`)+"\n")
+		return
+	}
+
+	if req.Method != "POST" {
+		w.WriteHeader(http.StatusNotFound)
+		return
+	}
+
+	username, password, ok := req.BasicAuth()
+	if !ok {
+		w.Header().Set("WWW-Authenticate", "Basic realm=\"travis-ci/worker\"")
+		w.WriteHeader(http.StatusUnauthorized)
+		return
+	}
+
+	authBytes := []byte(fmt.Sprintf("%s:%s", username, password))
+	if subtle.ConstantTimeCompare(authBytes, []byte(i.c.String("http-api-auth"))) != 1 {
+		w.WriteHeader(http.StatusForbidden)
+		return
+	}
+
+	w.Header().Set("Content-Type", "text/plain;charset=utf-8")
+
+	action := strings.ToLower(strings.Replace(req.URL.Path, "/worker/", "", 1))
+	i.logger.Info(fmt.Sprintf("web %s received", action))
+	switch action {
+	case "graceful-shutdown":
+		i.ProcessorPool.GracefulShutdown(false)
+		w.WriteHeader(http.StatusOK)
+		fmt.Fprintf(w, "starting graceful shutdown\n")
+	case "shutdown":
+		i.cancel()
+		fmt.Fprintf(w, "shutting down immediately\n")
+	case "pool-incr":
+		prev := i.ProcessorPool.Size()
+		i.ProcessorPool.Incr()
+		fmt.Fprintf(w, "adding processor to pool (%v + 1)\n", prev)
+	case "pool-decr":
+		prev := i.ProcessorPool.Size()
+		i.ProcessorPool.Decr()
+		fmt.Fprintf(w, "removing processor from pool (%v - 1)\n", prev)
+	case "graceful-shutdown-pause":
+		i.ProcessorPool.GracefulShutdown(true)
+		fmt.Fprintf(w, "toggling graceful shutdown and pause\n")
+	case "info":
+		fmt.Fprintf(w, "version: %s\n"+
+			"revision: %s\n"+
+			"generated: %s\n"+
+			"boot_time: %s\n"+
+			"uptime: %v\n"+
+			"pool_size: %v\n"+
+			"total_processed: %v\n"+
+			"processors:\n",
+			VersionString,
+			RevisionString,
+			GeneratedString,
+			i.bootTime.String(),
+			time.Since(i.bootTime),
+			i.ProcessorPool.Size(),
+			i.ProcessorPool.TotalProcessed())
+		i.ProcessorPool.Each(func(n int, proc *Processor) {
+			fmt.Fprintf(w, "- n: %v\n"+
+				"  id: %v\n"+
+				"  processed: %v\n"+
+				"  status: %v\n"+
+				"  last_job_id: %v\n",
+				n,
+				proc.ID,
+				proc.ProcessedCount,
+				proc.CurrentStatus,
+				proc.LastJobID)
+		})
+	default:
+		w.Header().Set("Travis-Worker-Unknown-Action", action)
+		w.WriteHeader(http.StatusNotFound)
+	}
 }
 
 func (i *CLI) signalHandler() {